+++ CHANGELOG.md
@@ -60,6 +60,7 @@
   This breaks for anybody who has already used `Polygon.clip_out_of_image()`. 
 * Fixed `Affine` heatmap augmentation producing arrays with values outside the range `[0.0, 1.0]` when `order` was set to `3`.
 * Changed `Affine` to always use `order=3` for heatmap augmentation.
+* Fixed `PiecewiseAffine` heatmap augmentation producing arrays with values outside the range `[0.0, 1.0]` when `order` was set to `3`.
 
 
 # 0.2.8